requirements.txt
insightface==0.7.3
pyheif==0.7.1
pillow==10.1.0
imageio==2.31.1
sklearn==1.3.1
numpy==1.24.4